LPS-60884 only delete toRepository folder when the new folder is empty

+++ portal-impl/src/com/liferay/portlet/documentlibrary/service/impl/DLAppServiceImpl.java
@@ -3317,7 +3317,12 @@ public class DLAppServiceImpl extends DLAppServiceBaseImpl {
 			}
 		}
 		catch (PortalException pe) {
-			toRepository.deleteFolder(newFolder.getFolderId());
+			if (getFoldersAndFileEntriesAndFileShortcutsCount(
+					newFolder.getRepositoryId(), newFolder.getFolderId(),
+					WorkflowConstants.STATUS_ANY, true) == 0) {
+
+				toRepository.deleteFolder(newFolder.getFolderId());
+			}
 
 			throw pe;
 		}
